add new model and new classification method

+++ src/algorithm/RangedOrderMC.hpp
@@ -178,7 +178,7 @@ namespace MC {
                 double p = probability( query.substr( 0, o ), query[o] );
                 acc += std::log( p );
             }
-            for (auto i = 0; i < query.size() - maxOrder() - 1; ++i)
+            for (auto i = 0; i < int64_t( query.size()) - maxOrder() - 1; ++i)
             {
                 double p = probability( query.substr( i, maxOrder()), query[i + maxOrder()] );
                 acc += std::log( p );
